fixes to stop time determination in GTFSDataModel.
use differences between repeated stop times on the same route direction to
calculate frequencies when not using frequencies.txt; do not assume there
exists a stop whose stop_sequence is 1.
consider a trip/segment to be running if it starts within one frequency
range of the current time. this fixes the problem where the stop list was
preferring times in the past.

in TransitDataModel, change error_log to Kurogo::log.
add functions to TransitTime to do addition/subtraction of seconds, and
allow negative components in createFromComponents.